Moved RedBlackTreeNode class to separate files
