+++ makefile
@@ -1,4 +1,4 @@
-a.out: main.o minalib.o addons.o addonsdb.o pokemon.o redblacktree.o kanto.o
+a.out: main.o minalib.o addons.o addonsdb.o redblacktreenode.o pokemon.o redblacktree.o kanto.o
 	g++ -g *.o
 
 main.o: main.cpp
@@ -12,6 +12,9 @@ redblacktree.o: redblacktree.h redblacktree.cpp
 
 pokemon.o: pokemonabc.h pokemonsub.h pokemonabc.cpp pokemonsub.cpp
 	g++ -g -c pokemonabc.cpp pokemonsub.cpp
+
+redblacktreenode.o: redblacktreenode.h redblacktreenode.cpp
+	g++ -g -c redblacktreenode.cpp
 
 addonsdb.o: addonsdb.h addonsdb.cpp
 	g++ -g -c addonsdb.cpp
